+++ crates/release_plz_core/src/command/mod.rs
@@ -1,6 +1,7 @@
 mod release;
 mod release_pr;
 pub mod set_version;
+mod trusted_publishing;
 mod update;
 
 pub use release::*;