+++ crates/test_logs/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.33](https://github.com/release-plz/release-plz/compare/test_logs-v0.1.32...test_logs-v0.1.33) - 2025-09-13
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.1.32](https://github.com/release-plz/release-plz/compare/test_logs-v0.1.31...test_logs-v0.1.32) - 2025-03-01
 
 ### Other